+++ snuba_sdk/metrics_visitors.py
@@ -5,7 +5,7 @@ from typing import Any, Generic, Mapping, TypeVar
 
 from snuba_sdk.column import Column
 from snuba_sdk.conditions import And, Condition, ConditionGroup, Op
-from snuba_sdk.expressions import InvalidExpressionError
+from snuba_sdk.expressions import InvalidExpressionError, Totals
 from snuba_sdk.function import CurriedFunction, Function
 from snuba_sdk.orderby import Direction, OrderBy
 from snuba_sdk.timeseries import Metric, MetricsScope, Rollup, Timeseries
@@ -159,14 +159,17 @@ class RollupSnQLPrinter(RollupVisitor[Mapping[str, str]]):
         self.translator = translator or Translation()
 
     def visit(self, rollup: Rollup) -> Mapping[str, str]:
-        condition = Condition(
-            lhs=Column("granularity"),
-            op=Op.EQ,
-            rhs=rollup.granularity,
-        )
+        condition = None
+        if rollup.granularity is not None:
+            condition = Condition(
+                lhs=Column("granularity"),
+                op=Op.EQ,
+                rhs=rollup.granularity,
+            )
 
         interval = ""
         orderby = ""
+        with_totals = ""
         if rollup.interval:
             interval_exp = Function(
                 "toStartOfInterval",
@@ -180,14 +183,17 @@ class RollupSnQLPrinter(RollupVisitor[Mapping[str, str]]):
             interval = self.translator.visit(interval_exp)
             orderby_exp = OrderBy(Column("time"), Direction.ASC)
             orderby = self.translator.visit(orderby_exp)
+            if rollup.totals:
+                with_totals = f"TOTALS {self.translator.visit(Totals(rollup.totals))}"
         elif rollup.orderby is not None:
             orderby_exp = OrderBy(Column(AGGREGATE_ALIAS), rollup.orderby)
             orderby = self.translator.visit(orderby_exp)
 
         return {
             "orderby": orderby,
-            "filter": self.translator.visit(condition),
+            "filter": self.translator.visit(condition) if condition else "",
             "interval": interval,
+            "with_totals": with_totals,
         }